+++ curie/calibration.py
@@ -170,7 +170,7 @@ class Calibration(object):
 	>>> cb.engcal = [0.1, 0.2, 0.003]
 	>>> print(cb.engcal)
 	[0.1   0.2   0.003]
-	>>> cb.saveas('test_calib.json')
+	>>> cb.saveas('test_calib.json')  # doctest: +SKIP
 	>>> cb = Calibration('test_calib.json')
 	>>> print(cb.engcal)
 	[0.1   0.2   0.003]