+++ curie/calibration.py
@@ -171,8 +171,8 @@ class Calibration(object):
 	>>> print(cb.engcal)
 	[0.1   0.2   0.003]
 	>>> cb.saveas('test_calib.json')  # doctest: +SKIP
-	>>> cb = Calibration('test_calib.json')
-	>>> print(cb.engcal)
+	>>> cb = Calibration('test_calib.json')  # doctest: +SKIP
+	>>> print(cb.engcal)  # doctest: +SKIP
 	[0.1   0.2   0.003]
 
 	"""